requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
kaleido==0.2.1
lxml==5.2.2
MarkupSafe==2.1.5
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
python-dateutil==2.9.0.post0
python-pptx==1.0.1
pytz==2024.1
six==1.16.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
XlsxWriter==3.2.0
